+++ functions/Add-PSBlueSkyImage.ps1
@@ -6,7 +6,7 @@ Function Add-BskyImage {
         [ValidateNotNullOrEmpty()]
         [ValidateScript({ Test-Path $_ },ErrorMessage = 'The file {0} could not be found.')]
         [ValidateScript({ (Get-Item $_).Length -lt 1MB }, ErrorMessage = 'The image file must be smaller than 1MB.')]
-        [ValidatePattern('.*\.(jpg|jpeg|png)$', ErrorMessage = 'The file must be a jpg, jpeg, or png file.')]
+        [ValidatePattern('.*\.(jpg|jpeg|png|gif)$', ErrorMessage = 'The file must be a jpg, jpeg, or png file.')]
         [Alias('Path')]
         [string]$ImagePath,
         [Parameter(HelpMessage = 'You should include ALT text for the image.')]